+++ lib/loadGenerators.js
